move extension contexts scripts
